+++ src/commons/utils/string.utils.ts
@@ -36,7 +36,33 @@ export class StringUtils {
         return keywordsList.join(',');
     }
 
-    static calcCharCodeForListIndex(code: number): string {
+    public static createReplacementsFromConfigArray(config: [any, any][]): [RegExp, string][] {
+        const replacementConfig = [];
+        if (Array.isArray(config)) {
+            for (const replacement of config) {
+                if (Array.isArray(replacement) && replacement.length === 2) {
+                    replacementConfig.push([new RegExp(replacement[0]), replacement[1]]);
+                }
+            }
+        }
+
+        return replacementConfig;
+    }
+
+    public static doReplacements(src: string, nameReplacements: [RegExp, string][]): string {
+        if (src === undefined || src === null || !nameReplacements || !Array.isArray(nameReplacements)) {
+            return src;
+        }
+
+        let name = src;
+        for (const replacement of nameReplacements) {
+            name = name.replace(replacement[0], replacement[1]);
+        }
+
+        return name;
+    }
+
+    public static calcCharCodeForListIndex(code: number): string {
         const baseChar = ('A').charCodeAt(0);
         let res  = '';
 
@@ -49,7 +75,7 @@ export class StringUtils {
         return res;
     }
 
-    static generateTechnicalName(name: string): string {
+    public static generateTechnicalName(name: string): string {
         return name ? name.replace(/[^-a-zA-Z0-9.+]+/g, ' ')
             .replace(/ +/g, ' ')
             .replace(/ /g, '-')
@@ -57,7 +83,7 @@ export class StringUtils {
             .toLowerCase() : '';
     }
 
-    static findNeedle(source: string, needle: string, findIdx: number): number {
+    public static findNeedle(source: string, needle: string, findIdx: number): number {
         let lastPos = -needle.length;
         let curPos = -1;
         let idx = -1;